requirements/base.txt
--extra-index-url=http://dist.pinaxproject.com/dev

setuptools
paver

# Extensions with C components
pillow
lxml

OWSLib==0.8.6
Django==1.5.5
django-extensions==0.8
httplib2==0.7.4
psycopg2==2.4.1
-e git+https://github.com/cga-harvard/gsconfig.py.git#egg=gsconfig.py
pinax-theme-bootstrap==3.0a11,
pinax-theme-bootstrap-account==1.0b2,
django-user-accounts==1.0b14,
django-forms-bootstrap==2.0.3.post1,
#-e git+https://github.com/mbertrand/django-avatar.git#egg=avatar
geonode-avatar==2.1.1
dialogos==0.1
agon-ratings==0.2
South==0.7.3
django-taggit==0.9.3
ordereddict
simplejson
defusedxml

#Uncomment the following to use memcached
#python-memcached

#The following are for the gazetteer
-e git+https://github.com/cga-harvard/geopy.git#egg=geopy
geojson
https://bitbucket.org/jespern/django-piston/get/5b3ef294fe89.zip


#For scheduled tasks
django-celery
kombu
pytz
#For gazetteer
python-dateutil==1.5 #For Python < 3.0
datautil
jdcal
#For mapnotes
vectorformats
#For autocomplete
django-autocomplete-light==1.4.5

unidecode

#ArcGIS REST services
-e git+https://github.com/Esri/arcpy-server-util-rest.git@hg/default#egg=arcrest

